rearranged device search by client name/product name/serial number into device modules
