+++ src/main/java/ee/valiit/project/Repository/RepositoryClient.java
@@ -44,20 +44,5 @@ public class RepositoryClient {
         return jdbcTemplate.query(sql, paramMap, new RowMapperClient());
     }
 
-    // devices by client, product or serialNr
-    public List<ClientEntity> getDevicesBy(String clientLike, String productLike, String serialNumberLike) {
-        String sql = "SELECT * FROM clients " +
-                "JOIN devices ON clients.id = devices.client_id " +
-                "JOIN products ON products.id = devices.product_id " +
-                "WHERE clients.client_name ILIKE :clientLike " +
-                "AND products.name ILIKE :productLike " +
-                "AND devices.sn ILIKE :serialNumberLike";
-        Map paramMap = new HashMap();
 
-        paramMap.put("clientLike", "%"+clientLike+"%");
-        paramMap.put("productLike", "%"+productLike+"%");
-        paramMap.put("serialNumberLike", "%"+serialNumberLike+"%");
-
-        return jdbcTemplate.query(sql, paramMap, new RowMapperDeviceMulti());
-    }
 }
