+++ src/main/java/ee/valiit/project/Service/ServiceProduct.java
@@ -1,9 +1,6 @@
 package ee.valiit.project.Service;
 
 import ee.valiit.project.Entity.ProductEntity;
-import ee.valiit.project.Entity.ProductEntity;
-import ee.valiit.project.Entity.Products;
-import ee.valiit.project.Repository.ProductRepository;
 import ee.valiit.project.Repository.RepositoryProduct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,13 +17,6 @@ public class ServiceProduct {
         repositoryProduct.newProduct(newProduct);
     }
 
-    //hibernate
-    public void newProductHibernate(String name, Integer stock){
-        Products products = new Products();
-        products.setName(name);
-        products.setStock(stock);
-        productRepository.save(products);
-    }
 
 
     public int productID(String name) {
@@ -38,10 +28,6 @@ public class ServiceProduct {
 
         return repositoryProduct.getProductInfo(id);
 
-    }
-    //hibernate get
-    public Products getProductHibernate(){
-        return productRepository.getOne(1);
     }
 
     //Get the whole list of Products
