+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.datasource.url=jdbc:postgresql://localhost/project
 spring.datasource.username=postgres
-spring.datasource.password=Parool11
+spring.datasource.password=1234
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.liquibase.change-log=classpath:/db/changelog/db.changelog-master.xml
